Pass CountData to descendants by hand for reading
It's easy for reading the value, but it requires passing the value by
hand to all the widgets that need it, which means all the widgets along
the way also need to get it to pass it further down. This doesn't scale,
results in unnecessary code as widgets that don't really need it know
about it, nor does it work for widgets we don't own the code for - we
can't just add a field for every thing that we want to pass down.

We could use some kinds of wrapper widgets but this gets ugly very
quickly.

Writing (incrementing in this case) is even more complex as we need a
way to update the UI, which doesn't work yet.