Make CounterWidget rebuild when value is written
This is done using StatefulWidget. But what with other widgets? We could
cause the whole UI to rebuild, but it is wasteful as it re-renders too
much.

Additionally, some widget along the way, above all widgets that need the
count, must also be stateful to initialize the state (App in our case).