Introduce InheritedWidget
Now only the widgets which really need the data can look it up, all
other are blissfully unaware of its existence.

There is normally always some kind of StatefulWidget of some kind above
an InheritedWidget so that there is a way to rebuild (parts of) the UI
and return a new instance of InheritedWidget for dependants to rebuilt
as well.

However, the code is a bit more complex now. Sadly, this is pretty much
the way it is done and means a lot of boilerplate.

Best example from Flutter is MediaQuery, which is an InheritedWidget an
it is managed by its own StatefulWidget in WidgetsApp.