Extract common CountText widget
NOTE: restart the application, don't just do hot-reload. The reason is
that some widgets will still be registered as dependants, their build()
will be triggered, and debug console output will be confusing and not
correctly show what is important.

This avoid codes duplication and allows to optimize what is rebuilt.