Use GlobalKey and break updating of deep descendant
NOTE: the application needs to be stopped and built and run again to see
that it is broken. I don't know why it is required, but I don't care
either ;d

Now one of the widgets doesn't depend on InheritedWidget and uses a
GlobalKey instead. This works only for reading data once, but it doesn't
register the widget as a dependency and Flutter doesn't automatically
rebuild it. Conclusion: GlobalKeys are not always a good way to look up
data.