Add client-server app
NOTE: uncomment the await on Future.delayed(100ms) in 'mapEventToState'
to really see that the UI is updated for each Success state. Without it,
the next InProgress state is triggered immediately after previous Success
and it is impossible to see UI changes. Without it, only looking at the
timestamps of the log output helps.

NOTE 2: 'asyncExpand' is called 'concatMap' in RX.

By default, events are processed sequentially and to completion, i.e. if
one event is being processed and takes some time and is not finished yet
but another event comes in, the new one needs to 'wait' until the
current one completes. This happens because by default 'transformEvents'
uses 'asyncExpand' ('concatMap' in RX parlance). This also means that by
default event processing cannot be cancelled mid way.

Click like mad to trigger many request very fast - the event will be
processed sequentially no matter what; it is slow, but safe as it is
kind of hard to break things.

See:
- https://github.com/felangel/bloc/issues/126
- https://rxjs.dev/api/operators/concatMap
- https://rxmarbles.com/#concatMap