Transform events with 'switchMap'
Now a new event interrupts processing of the previous one as soon as it
arrives, and takes over. This allows, for example, to cancel showing
stale (from previous event) server data. As soon as a new event arrives,
the stream resulting from the previous one, although not fully processed
yet, is cancelled and the new one is subscribed to.

In our use case, this results in the Success state being yielded only
for those events that have been fully processed, resulting in fewer UI
rebuilds.

Click like mad again and notice how it is all still sequential, but not
all states are yielded. The whole thing is still slow, though.

This use case is documented:
https://github.com/felangel/bloc/blob/bloc-v6.0.0/packages/bloc/lib/src/bloc.dart#L160-L168

See:
- https://github.com/felangel/bloc/issues/1238
- https://rxjs.dev/api/operators/switchMap
- https://rxmarbles.com/#switchMap