Transform events with 'flatMap'
NOTE: 'flatMap' is sometimes called 'mergeMap' in RX.

Now all events are processed mixed up, as soon as a new event comes it
will immediately start to have effect on the state. Now it is impossible
to foresee the ordering of states - 'flatMap' is able to subscribe to
multiple streams at the same time and yields whatever stream yields
whatever state.

If the states have a specific sequence and certain transitions are
invalid, 'flatMap' may introduce hardcore bugs if not taken care of
explicitly. mobile-connected doesn't use 'flatMap', but it does use
'switchMap' in a few places. It seems that 'flatMap' may not be a
popular operator for BLoC.

Click like mad - at some point an exception will be thrown due to a
state 'sanity check' in the 'mapEventToState' method. If such ordering
is important, a BLoC with 'flatMap' may be very complex and hard to
maintain.

See:
- https://rxjs.dev/api/operators/flatMap
- https://rxmarbles.com/#mergeMap
- https://tolikcode.github.io/post/rxjsMap/ for a comparison between
all the 'xxxMap' operators.