Debug why the UI needs so long with 'switchMap'
NOTE: to make the effect easier to note, make ColorServer slower by
restarting it and giving a bigger delay, like 3000.

Although 'switchMap' cancells the stream subscription and no unnecessary
states are yielded, it doesn't automatically take care or cancelling the
tasks already initiated. This needs to be programmed explicitly and also
supported by the underlying library.

In our case we see that even though new events come, HTTP tasks for old
ones are still excuted at the server side although the results will be
discarded anyway, which is wasteful.