Cancel outdated HTTP requests
Dart HttpClientRequest supports cancellation. However, the UI still
needs the same amount of time to show the correct result, and the reason
is that ColorServer is single-threaded rubbish.

Lesson: even though tutorials say 'switchMap' cancels the stream and all
is good it isn't necessarily true - more work may be necessary, e.g. the
underlying library and the server, if relevant, must play along.