Break the BLoC with 'switchMap'
NOTE: restart ColorServer and reduce the delay again.

Due to 'impure' 'mapEventToState' and its side effects, it is still
possible to break 'switchMap' (but not 'asyncExpand', at least not this
way, because it always processes the stream to completion).

Click like mad - at some point an error should be reported that array
lengths are inconsistent. This is due to the fact that streams can be
cancelled at 'yield' points, but between them all code will run. In our
case there are awaits which will allow the new stream 'switchMap'
switches execute its Futures, and change internal BLoC state; when the
stale streams resume, the checks they make cause errors.