Implement our Bloc
Implements a generic Bloc superclass and makes UsersBloc extend it.
Apart from a few minor modifications, the code is unchanged and pretty
simple, but broken - there is no way to propagate our Bloc down the
widget tree yet.

When the button is clicked, exceptions are thrown and logged in the
debug console bevause the BLoC cannot be found in the context, it needs
to be implemented.