Implement our BlocProvider
It is implemented in terms of Provider from the 'provider' package.

Now the BLoC can be propagates down the tree, but it doesn't react to
events - clicking the button doesn't trigger BLoC logic, and no events
are logged to the console at the server side.

The reason is that Streams, even broadcast, don't actually do anything
if they don't have any listeners subscribed, i.e. they are 'cold'.
(Single-subscription streams buffer events when there is no listener,
broadcast streams don't.)