Make our Bloc 'hot'
Now the events are processed as there is a subscriber (internal to the
BLoC, which adds each distinct state to the state controller).

As a result, even though there is no state listener yet, the BLoC still
does all the work internally and changes its state.

This can be verified by looking at the log output on the client and
server side.